完善了与插值程序的交互

+++ params.h
@@ -33,7 +33,7 @@ public:
     // IPL信息 // 不懂
     float interpPower;
     float radius;
-    float minPoint;
+    int minPoint;
 
     bool writeConfig();              // 写数据到 config.ini
